issue#6: rewrited ext lib in C insteadof C++.
