+++ src/_extdll/ext/fileio.c
@@ -1,12 +1,10 @@
 #include <kcc/dll.h>
 #include <kcc/dllcore.h>
-#include "../lib/fileio.hpp"
+#include "../lib/fileio.h"
 
 /* ---------------------------------------------------------------------------------------------
     FILE I/O
 --------------------------------------------------------------------------------------------- */
-
-extern "C" {
 
 DLLEXPORT void* fileio_get_iobuf(int argc, arg_type_t* argv)
 {
@@ -31,7 +29,7 @@ DLLEXPORT void* fileio_fopen(int argc, arg_type_t* argv)
     }
     const char *filename = (const char *)argv[0].value.s;
     const char *mode = (const char *)argv[1].value.s;
-    return klib::fopen(filename, mode);
+    return klib_fopen(filename, mode);
 }
 
 DLLEXPORT int fileio_fclose(int argc, arg_type_t* argv)
@@ -39,9 +37,9 @@ DLLEXPORT int fileio_fclose(int argc, arg_type_t* argv)
     if (argc != 1 || argv[0].type !=  C_PTR) {
         return EOF;
     }
-    klib::fileio *fp = (klib::fileio *)argv[0].value.p;
+    fileio *fp = (fileio *)argv[0].value.p;
     if (!fp) return EOF;
-    return klib::fclose(fp);
+    return klib_fclose(fp);
 }
 
 DLLEXPORT int fileio_fgetpos(int argc, arg_type_t* argv)
@@ -49,9 +47,9 @@ DLLEXPORT int fileio_fgetpos(int argc, arg_type_t* argv)
     if (argc != 2 || argv[0].type !=  C_PTR || argv[1].type !=  C_PTR) {
         return 0;
     }
-    klib::fileio *fp = (klib::fileio *)argv[0].value.p;
-    klib::fpos_t *pos = (klib::fpos_t *)argv[1].value.p;
-    return klib::fgetpos(fp, pos);
+    fileio *fp = (fileio *)argv[0].value.p;
+    klib_fpos_t *pos = (klib_fpos_t *)argv[1].value.p;
+    return klib_fgetpos(fp, pos);
 }
 
 DLLEXPORT int fileio_fsetpos(int argc, arg_type_t* argv)
@@ -59,9 +57,9 @@ DLLEXPORT int fileio_fsetpos(int argc, arg_type_t* argv)
     if (argc != 2 || argv[0].type !=  C_PTR || argv[1].type !=  C_PTR) {
         return 0;
     }
-    klib::fileio *fp = (klib::fileio *)argv[0].value.p;
-    klib::fpos_t *pos = (klib::fpos_t *)argv[1].value.p;
-    return klib::fsetpos(fp, pos);
+    fileio *fp = (fileio *)argv[0].value.p;
+    klib_fpos_t *pos = (klib_fpos_t *)argv[1].value.p;
+    return klib_fsetpos(fp, pos);
 }
 
 DLLEXPORT int fileio_fflush(int argc, arg_type_t* argv)
@@ -69,9 +67,9 @@ DLLEXPORT int fileio_fflush(int argc, arg_type_t* argv)
     if (argc != 1 || argv[0].type !=  C_PTR) {
         return 0;
     }
-    klib::fileio *fp = (klib::fileio *)argv[0].value.p;
+    fileio *fp = (fileio *)argv[0].value.p;
     if (!fp) return 0;
-    return klib::fflush(fp);
+    return klib_fflush(fp);
 }
 
 DLLEXPORT int fileio_feof(int argc, arg_type_t* argv)
@@ -79,9 +77,9 @@ DLLEXPORT int fileio_feof(int argc, arg_type_t* argv)
     if (argc != 1 || argv[0].type !=  C_PTR) {
         return 1;
     }
-    klib::fileio *fp = (klib::fileio *)argv[0].value.p;
+    fileio *fp = (fileio *)argv[0].value.p;
     if (!fp) return 1;
-    return klib::feof(fp);
+    return klib_feof(fp);
 }
 
 DLLEXPORT int fileio_fgetc(int argc, arg_type_t* argv)
@@ -89,9 +87,9 @@ DLLEXPORT int fileio_fgetc(int argc, arg_type_t* argv)
     if (argc != 1 || argv[0].type !=  C_PTR) {
         return 0;
     }
-    klib::fileio *fp = (klib::fileio *)argv[0].value.p;
+    fileio *fp = (fileio *)argv[0].value.p;
     if (!fp) return 0;
-    return klib::fgetc(fp);
+    return klib_fgetc(fp);
 }
 
 DLLEXPORT char *fileio_fgets(int argc, arg_type_t* argv)
@@ -101,9 +99,9 @@ DLLEXPORT char *fileio_fgets(int argc, arg_type_t* argv)
     }
     char *s = (char *)argv[0].value.s;
     int n = (int)argv[1].value.i;
-    klib::fileio *fp = (klib::fileio *)argv[2].value.p;
+    fileio *fp = (fileio *)argv[2].value.p;
     if (!fp) return 0;
-    return klib::fgets(s, n, fp);
+    return klib_fgets(s, n, fp);
 }
 
 DLLEXPORT int fileio_fputc(int argc, arg_type_t* argv)
@@ -112,9 +110,9 @@ DLLEXPORT int fileio_fputc(int argc, arg_type_t* argv)
         return 0;
     }
     int c = (int)argv[0].value.i;
-    klib::fileio *fp = (klib::fileio *)argv[1].value.p;
+    fileio *fp = (fileio *)argv[1].value.p;
     if (!fp) return 0;
-    return klib::fputc(c, fp);
+    return klib_fputc(c, fp);
 }
 
 DLLEXPORT int fileio_fputs(int argc, arg_type_t* argv)
@@ -123,9 +121,9 @@ DLLEXPORT int fileio_fputs(int argc, arg_type_t* argv)
         return 0;
     }
     const char *s = (const char *)argv[0].value.s;
-    klib::fileio *fp = (klib::fileio *)argv[1].value.p;
+    fileio *fp = (fileio *)argv[1].value.p;
     if (!fp) return 0;
-    return klib::fputs(s, fp);
+    return klib_fputs(s, fp);
 }
 
 DLLEXPORT unsigned int fileio_fread(int argc, arg_type_t* argv)
@@ -136,9 +134,9 @@ DLLEXPORT unsigned int fileio_fread(int argc, arg_type_t* argv)
     void *ptr = (void *)argv[0].value.p;
     size_t size = (int)argv[1].value.i;
     size_t nmemb = (int)argv[2].value.i;
-    klib::fileio *fp = (klib::fileio *)argv[3].value.p;
+    fileio *fp = (fileio *)argv[3].value.p;
     if (!fp) return 0;
-    return klib::fread(ptr, size, nmemb, fp);
+    return klib_fread(ptr, size, nmemb, fp);
 }
 
 DLLEXPORT unsigned int fileio_fwrite(int argc, arg_type_t* argv)
@@ -149,9 +147,9 @@ DLLEXPORT unsigned int fileio_fwrite(int argc, arg_type_t* argv)
     const void *ptr = (const void *)argv[0].value.p;
     size_t size = (int)argv[1].value.i;
     size_t nmemb = (int)argv[2].value.i;
-    klib::fileio *fp = (klib::fileio *)argv[3].value.p;
+    fileio *fp = (fileio *)argv[3].value.p;
     if (!fp) return 0;
-    return klib::fwrite(ptr, size, nmemb, fp);
+    return klib_fwrite(ptr, size, nmemb, fp);
 }
 
 DLLEXPORT int fileio_fseek(int argc, arg_type_t* argv)
@@ -159,11 +157,11 @@ DLLEXPORT int fileio_fseek(int argc, arg_type_t* argv)
     if (argc != 3 || argv[0].type !=  C_PTR || argv[1].type !=  C_INT || argv[2].type !=  C_INT) {
         return 0;
     }
-    klib::fileio *fp = (klib::fileio *)argv[0].value.p;
+    fileio *fp = (fileio *)argv[0].value.p;
     if (!fp) return 0;
     int64_t offset = (int64_t)argv[1].value.i;
     int whence = (int)argv[2].value.i;
-    return klib::fseek(fp, offset, whence);
+    return klib_fseek(fp, offset, whence);
 }
 
 DLLEXPORT int64_t fileio_ftell(int argc, arg_type_t* argv)
@@ -171,9 +169,9 @@ DLLEXPORT int64_t fileio_ftell(int argc, arg_type_t* argv)
     if (argc != 1 || argv[0].type !=  C_PTR) {
         return 0;
     }
-    klib::fileio *fp = (klib::fileio *)argv[0].value.p;
+    fileio *fp = (fileio *)argv[0].value.p;
     if (!fp) return 0;
-    return klib::ftell(fp);
+    return klib_ftell(fp);
 }
 
 DLLEXPORT void fileio_rewind(int argc, arg_type_t* argv)
@@ -181,9 +179,9 @@ DLLEXPORT void fileio_rewind(int argc, arg_type_t* argv)
     if (argc != 1 || argv[0].type !=  C_PTR) {
         return;
     }
-    klib::fileio *fp = (klib::fileio *)argv[0].value.p;
+    fileio *fp = (fileio *)argv[0].value.p;
     if (!fp) return;
-    klib::rewind(fp);
+    klib_rewind(fp);
 }
 
 DLLEXPORT int fileio_ungetc(int argc, arg_type_t* argv)
@@ -192,9 +190,9 @@ DLLEXPORT int fileio_ungetc(int argc, arg_type_t* argv)
         return 0;
     }
     int c = (int)argv[0].value.i;
-    klib::fileio *fp = (klib::fileio *)argv[1].value.p;
+    fileio *fp = (fileio *)argv[1].value.p;
     if (!fp) return 0;
-    return klib::ungetc(c, fp);
+    return klib_ungetc(c, fp);
 }
 
 DLLEXPORT int fileio_printf_ld(int argc, arg_type_t* argv)
@@ -202,11 +200,11 @@ DLLEXPORT int fileio_printf_ld(int argc, arg_type_t* argv)
     if (argc != 3 || argv[0].type !=  C_PTR || argv[1].type !=  C_STR || argv[2].type !=  C_INT) {
         return 0;
     }
-    klib::fileio *fp = (klib::fileio *)argv[0].value.p;
+    fileio *fp = (fileio *)argv[0].value.p;
     if (!fp) return 0;
     const char *fmt = (const char *)argv[1].value.s;
     long v = (long)argv[2].value.i;
-    return klib::fprintf(fp, fmt, v);
+    return klib_fprintf(fp, fmt, v);
 }
 
 DLLEXPORT int fileio_printf_d(int argc, arg_type_t* argv)
@@ -214,11 +212,11 @@ DLLEXPORT int fileio_printf_d(int argc, arg_type_t* argv)
     if (argc != 3 || argv[0].type !=  C_PTR || argv[1].type !=  C_STR || argv[2].type !=  C_INT) {
         return 0;
     }
-    klib::fileio *fp = (klib::fileio *)argv[0].value.p;
+    fileio *fp = (fileio *)argv[0].value.p;
     if (!fp) return 0;
     const char *fmt = (const char *)argv[1].value.s;
     int v = (int)argv[2].value.i;
-    return klib::fprintf(fp, fmt, v);
+    return klib_fprintf(fp, fmt, v);
 }
 
 DLLEXPORT int fileio_printf_lf(int argc, arg_type_t* argv)
@@ -226,11 +224,11 @@ DLLEXPORT int fileio_printf_lf(int argc, arg_type_t* argv)
     if (argc != 3 || argv[0].type !=  C_PTR || argv[1].type !=  C_STR || argv[2].type !=  C_DBL) {
         return 0;
     }
-    klib::fileio *fp = (klib::fileio *)argv[0].value.p;
+    fileio *fp = (fileio *)argv[0].value.p;
     if (!fp) return 0;
     const char *fmt = (const char *)argv[1].value.s;
     long double v = (long double)argv[2].value.d;
-    return klib::fprintf(fp, fmt, v);
+    return klib_fprintf(fp, fmt, v);
 }
 
 DLLEXPORT int fileio_printf_f(int argc, arg_type_t* argv)
@@ -238,11 +236,11 @@ DLLEXPORT int fileio_printf_f(int argc, arg_type_t* argv)
     if (argc != 3 || argv[0].type !=  C_PTR || argv[1].type !=  C_STR || argv[2].type !=  C_DBL) {
         return 0;
     }
-    klib::fileio *fp = (klib::fileio *)argv[0].value.p;
+    fileio *fp = (fileio *)argv[0].value.p;
     if (!fp) return 0;
     const char *fmt = (const char *)argv[1].value.s;
     double v = (double)argv[2].value.d;
-    return klib::fprintf(fp, fmt, v);
+    return klib_fprintf(fp, fmt, v);
 }
 
 DLLEXPORT int fileio_printf_p(int argc, arg_type_t* argv)
@@ -250,11 +248,11 @@ DLLEXPORT int fileio_printf_p(int argc, arg_type_t* argv)
     if (argc != 3 || argv[0].type !=  C_PTR || argv[1].type !=  C_STR || argv[2].type !=  C_PTR) {
         return 0;
     }
-    klib::fileio *fp = (klib::fileio *)argv[0].value.p;
+    fileio *fp = (fileio *)argv[0].value.p;
     if (!fp) return 0;
     const char *fmt = (const char *)argv[1].value.s;
     void *v = (void *)argv[2].value.p;
-    return klib::fprintf(fp, fmt, v);
+    return klib_fprintf(fp, fmt, v);
 }
 
 DLLEXPORT int fileio_printf_s(int argc, arg_type_t* argv)
@@ -262,11 +260,9 @@ DLLEXPORT int fileio_printf_s(int argc, arg_type_t* argv)
     if (argc != 3 || argv[0].type !=  C_PTR || argv[1].type !=  C_STR || argv[2].type !=  C_STR) {
         return 0;
     }
-    klib::fileio *fp = (klib::fileio *)argv[0].value.p;
+    fileio *fp = (fileio *)argv[0].value.p;
     if (!fp) return 0;
     const char *fmt = (const char *)argv[1].value.s;
     char *v = (char *)argv[2].value.s;
-    return klib::fprintf(fp, fmt, v);
+    return klib_fprintf(fp, fmt, v);
 }
-
-} // extern "C"
